Moved Helper To Lib Dir and Props To Types
